LayoutTests/svg/dynamic-updates/SVGFETurbulenceElement-svgdom-numOctaves-prop-expected.txt

CONSOLE MESSAGE: line 37: TypeError: 'undefined' is not an object (evaluating 'turbulence.numOctaves.baseVal = 3')
FAIL: Timed out waiting for notifyDone to be called
SVG 1.1 dynamic update tests

Tests dynamic updates of the 'numOctaves' property of the SVGFETurbulenceElment object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL turbulence.numOctaves.baseVal should be 1. Threw exception TypeError: 'undefined' is not an object (evaluating 'turbulence.numOctaves.baseVal')
